pull staff only views into staff app
