+++ jobserver/templates/backend_edit.html
@@ -8,7 +8,7 @@
     <li class="breadcrumb-item"><a href="/">Home</a></li>
 
     <li class="breadcrumb-item">
-      <a href="{% url 'backend-list' %}">Backends</a>
+      <a href="{% url 'staff:backend-list' %}">Backends</a>
     </li>
 
     <li class="breadcrumb-item">
