+++ jobserver/templates/user_detail.html
@@ -11,7 +11,7 @@
     <li class="breadcrumb-item"><a href="/">Home</a></li>
 
     <li class="breadcrumb-item">
-      <a href="{% url 'user-list' %}">Users</a>
+      <a href="{% url 'staff:user-list' %}">Users</a>
     </li>
 
     <li class="breadcrumb-item active" aria-current="page">{{ user.username }}</li>
