+++ jobserver/templates/user_list.html
@@ -48,7 +48,7 @@
 
     {% if request.GET %}
     <div class="mb-3">
-      <a href="{% url 'user-list' %}">Clear All</a>
+      <a href="{% url 'staff:user-list' %}">Clear All</a>
     </div>
     {% endif %}
