+++ staff/forms.py
@@ -1,0 +1,25 @@
+from django import forms
+
+from jobserver.authorization.forms import RolesForm
+from jobserver.backends import backends_to_choices
+from jobserver.models import Backend
+
+
+class UserForm(RolesForm):
+    is_superuser = forms.BooleanField(required=False)
+
+    def __init__(self, *, available_backends, **kwargs):
+        super().__init__(**kwargs)
+
+        # build choices from the available backends
+        choices = backends_to_choices(available_backends)
+
+        self.fields["backends"] = forms.MultipleChoiceField(
+            choices=choices,
+            required=False,
+            widget=forms.CheckboxSelectMultiple,
+        )
+
+    def clean_backends(self):
+        """Convert backend names to Backend instances"""
+        return Backend.objects.filter(slug__in=self.cleaned_data["backends"])
