+++ staff/urls.py
@@ -1,3 +1,28 @@
+from django.urls import include, path
+
+from .views.backends import BackendDetail, BackendEdit, BackendList, BackendRotateToken
+from .views.users import UserDetail, UserList
+
+
 app_name = "staff"
 
-urlpatterns = []
+backend_urls = [
+    path("", BackendList.as_view(), name="backend-list"),
+    path("<int:pk>/", BackendDetail.as_view(), name="backend-detail"),
+    path("<int:pk>/edit/", BackendEdit.as_view(), name="backend-edit"),
+    path(
+        "<int:pk>/rotate-token/",
+        BackendRotateToken.as_view(),
+        name="backend-rotate-token",
+    ),
+]
+
+user_urls = [
+    path("", UserList.as_view(), name="user-list"),
+    path("<username>/", UserDetail.as_view(), name="user-detail"),
+]
+
+urlpatterns = [
+    path("backends/", include(backend_urls)),
+    path("users/", include(user_urls)),
+]
