+++ tests/unit/jobserver/models/test_backends.py
@@ -28,7 +28,7 @@ def test_backend_get_absolute_url():
 
     url = backend.get_absolute_url()
 
-    assert url == reverse("backend-detail", kwargs={"pk": backend.pk})
+    assert url == reverse("staff:backend-detail", kwargs={"pk": backend.pk})
 
 
 @pytest.mark.django_db
@@ -37,7 +37,7 @@ def test_backend_get_edit_url():
 
     url = backend.get_edit_url()
 
-    assert url == reverse("backend-edit", kwargs={"pk": backend.pk})
+    assert url == reverse("staff:backend-edit", kwargs={"pk": backend.pk})
 
 
 @pytest.mark.django_db
@@ -46,7 +46,7 @@ def test_backend_get_rotate_url():
 
     url = backend.get_rotate_url()
 
-    assert url == reverse("backend-rotate-token", kwargs={"pk": backend.pk})
+    assert url == reverse("staff:backend-rotate-token", kwargs={"pk": backend.pk})
 
 
 @pytest.mark.django_db
