+++ tests/unit/jobserver/test_urls.py
@@ -24,7 +24,6 @@ from jobserver.api.releases import (
 from jobserver.utils import dotted_path
 from jobserver.views import (
     admin,
-    backends,
     index,
     job_requests,
     jobs,
@@ -35,6 +34,8 @@ from jobserver.views import (
     users,
     workspaces,
 )
+from staff.views import backends as staff_backends
+from staff.views import users as staff_users
 
 
 @pytest.mark.parametrize(
@@ -71,12 +72,14 @@ def test_url_redirects(client, url, redirect):
         ("/api/v2/releases/workspace/w", ReleaseWorkspaceAPI),
         ("/api/v2/releases/release/42", ReleaseAPI),
         ("/api/v2/releases/file/42", ReleaseFileAPI),
-        ("/backends/", backends.BackendList),
-        ("/backends/42/", backends.BackendDetail),
-        ("/backends/42/edit/", backends.BackendEdit),
-        ("/backends/42/rotate-token/", backends.BackendRotateToken),
         ("/event-log/", job_requests.JobRequestList),
-        ("/job-requests/<pk>/", job_requests.JobRequestDetailRedirect),
+        ("/staff/backends/", staff_backends.BackendList),
+        ("/staff/backends/42/", staff_backends.BackendDetail),
+        ("/staff/backends/42/edit/", staff_backends.BackendEdit),
+        ("/staff/backends/42/rotate-token/", staff_backends.BackendRotateToken),
+        ("/staff/users/", staff_users.UserList),
+        ("/staff/users/<username>/", staff_users.UserDetail),
+        ("/job-requests/42/", job_requests.JobRequestDetailRedirect),
         ("/jobs/<identifier>/", jobs.JobDetailRedirect),
         ("/login/github/", social_django_auth_view),
         ("/logout/", LogoutView),
@@ -84,8 +87,6 @@ def test_url_redirects(client, url, redirect):
         ("/orgs/new/", orgs.OrgCreate),
         ("/settings/", users.Settings),
         ("/status/", status.Status),
-        ("/users/", users.UserList),
-        ("/users/ben/", users.UserDetail),
         ("/o/", orgs.OrgDetail),
         ("/o/new-project/", projects.ProjectCreate),
         ("/o/project-onboarding/", projects.ProjectOnboardingCreate),
@@ -130,7 +131,7 @@ def test_url_resolution(url, view):
     """Test each URL resolves to the expected view function or class"""
     resolved_view = resolve(url).func
 
-    assert dotted_path(resolved_view) == dotted_path(view)
+    assert dotted_path(resolved_view) == dotted_path(view), url
 
     msg = f"Resolved view '{resolved_view}' is a class. Did you forget `.as_view()`?"
     assert not inspect.isclass(resolved_view), msg
